Vlib: sync write fns, fix for #2583 (backport from trunk)
git-svn-id: https://svn.osgeo.org/grass/grass/branches/releasebranch_7_0@64654 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ include/defs/vector.h
@@ -539,11 +539,11 @@ int V2_read_next_line_pg(struct Map_info *, struct line_pnts *,
 int V1_delete_line_nat(struct Map_info *, off_t);
 int V1_delete_line_ogr(struct Map_info *, off_t);
 int V1_delete_line_pg(struct Map_info *, off_t);
-int V2_delete_line_nat(struct Map_info *, int);
-int V2_delete_line_sfa(struct Map_info *, int);
-int V2_delete_line_pg(struct Map_info *, int);
+int V2_delete_line_nat(struct Map_info *, off_t);
+int V2_delete_line_sfa(struct Map_info *, off_t);
+int V2_delete_line_pg(struct Map_info *, off_t);
 int V1_restore_line_nat(struct Map_info *, off_t);
-int V2_restore_line_nat(struct Map_info *, int);
+int V2_restore_line_nat(struct Map_info *, off_t);
 off_t V1_write_line_nat(struct Map_info *, int, const struct line_pnts *,
                         const struct line_cats *);
 off_t V1_write_line_ogr(struct Map_info *, int, const struct line_pnts *,
@@ -562,11 +562,11 @@ off_t V1_rewrite_line_ogr(struct Map_info *, off_t, int,
                           const struct line_pnts *, const struct line_cats *);
 off_t V1_rewrite_line_pg(struct Map_info *, off_t, int,
                          const struct line_pnts *, const struct line_cats *);
-off_t V2_rewrite_line_nat(struct Map_info *, int, int,
+off_t V2_rewrite_line_nat(struct Map_info *, off_t, int,
                           const struct line_pnts *, const struct line_cats *);
-off_t V2_rewrite_line_sfa(struct Map_info *, int, int,
+off_t V2_rewrite_line_sfa(struct Map_info *, off_t, int,
                           const struct line_pnts *, const struct line_cats *);
-off_t V2_rewrite_line_pg(struct Map_info *, int, int,
+off_t V2_rewrite_line_pg(struct Map_info *, off_t, int,
                           const struct line_pnts *, const struct line_cats *);
 
     /* Build topology */
